+++ modules/likemode_superlike.js
@@ -9,3 +9,12 @@
  * @changelog:  0.1 initial release
  *
  */
+class Likemode_superlike {
+    constructor(bot, config, utils) {
+        this.bot = bot;
+        this.config = config;
+        this.utils = utils;
+    }
+}
+
+module.exports = (bot, config, utils) => { return new Likemode_superlike(bot, config, utils); };