+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+# v0.2.0 (January 15, 2018)
+
+* New flow pattern
+* Refactor: removed eval()
+* Refactor: removed setTimeout() and added system-sleep (run: npm install)
+* Refactor: folder routes moved to modules 
+* Feature: switch_mode()
+* Fix: 2FA bugs
+* Fix: errors manager bugs
+* Fix: screenshot bugs
+* Fix: random dislike
+
+
+# v0.1.1 (January 13, 2018)
+
+* Fix debug (verbose)
+* Updated README.md
+
+
 # v0.1.0 (January 12, 2018)
 
 * First Release.
+* Login
+* 2FA
+* Multi-Session
+* Errors manager (bad pin, bad password)
+* Screenshot and Verbose logger
+* Like Mode Classic: bot select random hashtag from config list and like 1 random photo (of last 20), and repeat this all time | 750-900 like/day. Limit is 1000/day for ig. This is safe mode.